adaptable cycle time
this is an experiment. Most likely i will revert this and reflect a shorter cycle time by adjusting the transition probabilities and inputs outside the model (as pre-model analysis). That is more simple, easier to implement in spreadsheet version and more transparent as it forces to think about how to adjust the various transition probabilities (separately or simultaniously using eigenvector)