simplified coding structure
Simplified the coding structure by moving calculations to prepare for strategy to the f.run_strategy function. That way, there is no longer a need to store the pre-calculations to the inputs list and bring them to the next function. The f.run_scenario function is only used for wrapping up the results (initializing and storing the strategy and scenario results).